Use size_t instead of int so both sides of a comparison are unsigned.

+++ tools/TextTool2/TextTool2.cpp
@@ -143,8 +143,8 @@ int main(int argc, char* argv[])
   // verify no files would be overwritten in output directory
   std::vector<std::string> outputFileList;
 
-  for(int i = 0; i < inputFonts.size(); ++i) {
-    for(int p = 0; p < (sizeof fontSizes / sizeof(const char *)); ++p) {
+  for(size_t i = 0; i < inputFonts.size(); ++i) {
+    for(size_t p = 0; p < (sizeof fontSizes / sizeof(const char *)); ++p) {
       outputFileList.push_back(stripHyphen(inputFonts[i]) + "_" + fontSizes[p] + ".png");
       outputFileList.push_back(stripHyphen(inputFonts[i]) + "_" + fontSizes[p] + ".fmt");
     }
@@ -155,7 +155,7 @@ int main(int argc, char* argv[])
   while((directoryEntry = readdir(outputDirectoryStream))) {
     std::string fileItem = directoryEntry->d_name;
 
-    for(int i = 0; i < outputFileList.size(); ++i) {
+    for(size_t i = 0; i < outputFileList.size(); ++i) {
       if(outputFileList[i] == fileItem) {
 	std::cerr <<
 	    "Directory " <<
@@ -187,7 +187,7 @@ int main(int argc, char* argv[])
 
   std::vector<FT_Face> ftFaces;
 
-  for(int i = 0; i < inputFonts.size(); ++i) {
+  for(size_t i = 0; i < inputFonts.size(); ++i) {
     std::string fontFilePath = inputDirectory + "/" + inputFonts[i] + ".ttf";
 
     std::cerr << "Loading font file " << fontFilePath << "... ";
@@ -208,8 +208,8 @@ int main(int argc, char* argv[])
   std::cerr << std::endl;
 
   // render textures and write files
-  for(int i = 0; i < inputFonts.size(); ++i) {
-    for(int p = 0; p < (sizeof fontSizes / sizeof(const char *)); ++p) {
+  for(size_t i = 0; i < inputFonts.size(); ++i) {
+    for(size_t p = 0; p < (sizeof fontSizes / sizeof(const char *)); ++p) {
       std::string fontName = stripHyphen(inputFonts[i]) + "_" + fontSizes[p];
       int fontSize = atoi(fontSizes[p]);
 
@@ -226,7 +226,7 @@ int main(int argc, char* argv[])
 
       std::vector<struct CharInfo> charInfoList;
 
-      for(int j = 0; j < strlen(requiredCharacters); ++j) {
+      for(size_t j = 0; j < strlen(requiredCharacters); ++j) {
 	if((ftError = FT_Load_Glyph(ftFaces[i], FT_Get_Char_Index(ftFaces[i], requiredCharacters[j]), FT_LOAD_DEFAULT))) {
 	  std::cerr <<
 	      "failed; unable to load glyph for character '" <<
@@ -285,7 +285,7 @@ int main(int argc, char* argv[])
       // create and initialize the texture
       int maxPixels = (int) ftFaces[i]->size->metrics.height / 64; // line height
 
-      for(int j = 0; j < charInfoList.size(); ++j)
+      for(size_t j = 0; j < charInfoList.size(); ++j)
 	if(charInfoList[j].width > maxPixels)
 	  maxPixels = charInfoList[j].width;
 
@@ -315,7 +315,7 @@ int main(int argc, char* argv[])
 	std::cerr << "failed; unable to write font format file " <<
 	    outputDirectory << "/" << fontName << ".fmt." << std::endl;
 
-	for(int j = 0; j < charInfoList.size(); ++j)
+	for(size_t j = 0; j < charInfoList.size(); ++j)
 	  free(charInfoList[j].bitmap);
 
 	return EXIT_FAILURE;
@@ -329,7 +329,7 @@ int main(int argc, char* argv[])
 	  "\n";
 
       // copy all of the character bitmap values into the texture's alpha channel and write format data
-      for(int j = 0; j < charInfoList.size(); ++j) {
+      for(size_t j = 0; j < charInfoList.size(); ++j) {
 	int basePositionX = j % (textureWidth / maxPixels) * maxPixels + charInfoList[j].adjustX;
 	int basePositionY = j / (textureWidth / maxPixels) * maxPixels +
 	    (int) ftFaces[i]->size->metrics.ascender / 64 -
@@ -380,7 +380,7 @@ int main(int argc, char* argv[])
 
 	free(texture);
 
-	for(int j = 0; j < charInfoList.size(); ++j)
+	for(size_t j = 0; j < charInfoList.size(); ++j)
 	  free(charInfoList[j].bitmap);
 
 	return EXIT_FAILURE;
@@ -391,7 +391,7 @@ int main(int argc, char* argv[])
       free(texture);
 
       // release character bitmap memory
-      for(int j = 0; j < charInfoList.size(); ++j)
+      for(size_t j = 0; j < charInfoList.size(); ++j)
 	free(charInfoList[j].bitmap);
 
       std::cerr << "done." << std::endl;
